refactor: subpath 추가 및 utils 폴더 묶음
