+++ src/app/reservation/[id]/write/utils/calculateDDayOnBooking.util.ts
